Course Timetabling Solver: Restore
- fixed a NumberFormatException when restoring conflict-based statistics and there is a constraint with a long id (cannot be parsed as integer)

+++ JavaSource/org/unitime/timetable/solver/ui/ConflictStatisticsInfo.java
@@ -452,7 +452,7 @@ public class ConflictStatisticsInfo implements TimetableInfo, Serializable {
 		}
 		CBSConstraint(CBSValue value, Element element) {
 			iValue = value;
-			iId = Integer.parseInt(element.attributeValue("id"));
+			iId = Long.parseLong(element.attributeValue("id"));
 			iType = Integer.parseInt(element.attributeValue("type"));
 			iName = element.attributeValue("name");
 			iPref = element.attributeValue("pref");
